Allow options to pass to UploadRequest.new
If we want to upload a randomized filename from a uuid, but have
the browser able to download the attachment with the real filename
we need to be able to pass an attachment filename for content
disposition use.

Also note that later these options can be used to pass aws acl
if we want to upload private files.